Separating the leaders page from the Our story page
